refactor: improve encapsulation
